Add copyrights and licenses
Add copyright notices to all files except for the Cargo build files, add
licenses for third party rust dependencies, and add a separate MIT0
License specifically for the provided example. This will allow anyone to
freely use the sample provided in the examples folder in their own
projects. The rest of the code remains under the Apache 2.0 License.

Signed-off-by: Connor Williamson <connordw@amazon.com>
Reviewed-by: Costin Lupu <lvpv@amazon.nl>
Reviewed-by: Mark Kirichenko <mkirich@amazon.de>
Reviewed-by: Marius Knaust <mknaust@amazon.de>